+++ HomeAccessPlus/AppDelegate.swift
@@ -20,6 +20,29 @@
 //
 
 import UIKit
+import XCGLogger
+
+// Declaring a global constant to the default XCGLogger instance
+let logger = XCGLogger.defaultInstance()
+
+// Declaring a global constant to the HAP+ main app colour
+let hapMainColour = "#005DAB"
+
+// Declaring a global constant to NSUserDefaults to access settings
+// throughout the app, and also the variables to these settings, to
+// prevent accidental typos with incorrect names
+// See: http://www.codingexplorer.com/nsuserdefaults-a-swift-introduction/
+let settings = NSUserDefaults.standardUserDefaults()
+let settingsHAPServer = "hapServer"
+let settingsSiteName = "siteName"
+let settingsFirstName = "firstName"
+let settingsUsername = "username"
+let settingsPassword = "password"
+let settingsToken1 = "token1"
+let settingsToken2 = "token2"
+let settingsToken2Name = "token2Name"
+let settingsDeviceType = "deviceType"
+let settingsUserRoles = "userRoles"
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate, UISplitViewControllerDelegate {
@@ -29,6 +52,10 @@ class AppDelegate: UIResponder, UIApplicationDelegate, UISplitViewControllerDele
 
     func application(application: UIApplication, didFinishLaunchingWithOptions launchOptions: [NSObject: AnyObject]?) -> Bool {
         // Override point for customization after application launch.
+        
+        // Configuring the logger options
+        logger.setup(.Debug, showThreadName: true, showLogLevel: true, showFileNames: true, showLineNumbers: true, showDate: true, writeToFile: nil, fileLogLevel: .Debug)
+        
         return true
     }