+++ Pods/Target Support Files/Pods-HomeAccessPlus/Pods-HomeAccessPlus-frameworks.sh
@@ -86,8 +86,16 @@ strip_invalid_archs() {
 if [[ "$CONFIGURATION" == "Debug" ]]; then
   install_framework "Pods-HomeAccessPlus/Alamofire.framework"
   install_framework "Pods-HomeAccessPlus/ChameleonFramework.framework"
+  install_framework "Pods-HomeAccessPlus/Locksmith.framework"
+  install_framework "Pods-HomeAccessPlus/MBProgressHUD.framework"
+  install_framework "Pods-HomeAccessPlus/SwiftyJSON.framework"
+  install_framework "Pods-HomeAccessPlus/XCGLogger.framework"
 fi
 if [[ "$CONFIGURATION" == "Release" ]]; then
   install_framework "Pods-HomeAccessPlus/Alamofire.framework"
   install_framework "Pods-HomeAccessPlus/ChameleonFramework.framework"
+  install_framework "Pods-HomeAccessPlus/Locksmith.framework"
+  install_framework "Pods-HomeAccessPlus/MBProgressHUD.framework"
+  install_framework "Pods-HomeAccessPlus/SwiftyJSON.framework"
+  install_framework "Pods-HomeAccessPlus/XCGLogger.framework"
 fi